small readme update + add new setting to choose to open docs in webview or browser
